+++ GoTrack/app/src/main/java/de/mobcom/group3/gotrack/Database/Models/Route.java
@@ -8,14 +8,14 @@ public class Route {
     private int id;
     private int userId;
     private String name;
-    private double time;
+    private long time;
     private long rideTime;
     private double distance;
     private ArrayList<Location> locations;
 
     public Route() {}
 
-    public Route(int id, int userId, String name, double time, long rideTime, double distance, ArrayList<Location> locations) {
+    public Route(int id, int userId, String name, long time, long rideTime, double distance, ArrayList<Location> locations) {
         this.id = id;
         this.userId = userId;
         this.name = name;
@@ -25,7 +25,7 @@ public class Route {
         this.locations =locations;
     }
 
-    public Route(int userId, String name, double time, long rideTime, double distance, ArrayList<Location> locations) {
+    public Route(int userId, String name, long time, long rideTime, double distance, ArrayList<Location> locations) {
         this.userId = userId;
         this.name = name;
         this.time = time;
@@ -57,10 +57,10 @@ public class Route {
         this.name = name;
     }
 
-    public double getTime() {
+    public long getTime() {
         return time;
     }
-    public void setTime(double time) {
+    public void setTime(long time) {
         this.time = time;
     }
